refactor(agents): refactor agents.py into a package agents with the modules base_agents and signal_agents
